requirements.txt
# Core framework
streamlit>=1.32.0

# LangChain ecosystem
langchain>=0.1.16
langchain-community>=0.0.38
langchain-openai>=0.1.1
langchain-core>=0.1.42
langchain-ollama>=0.1.0
langchain-groq>=0.1.3

# Vector store
chromadb>=0.4.24

# LLM clients
openai>=1.14.0
ollama>=0.2.0

# OCR and image processing
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.3.0
opencv-python-headless>=4.9.0

# PDF generation
reportlab>=4.1.0

# PDF reading (PyMuPDF)
pymupdf>=1.24.0

# Data
pydantic>=2.6.0
python-dotenv>=1.0.1
pandas>=2.2.1
numpy>=1.26.4

# PII detection (fully local — no cloud)
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354

# NLP
nltk>=3.8.1
rapidfuzz>=3.6.1

# NOTE: Do NOT install spacy — not needed, saves install time
# NOTE: Do NOT install sentence-transformers — use Ollama gte-large instead
# NOTE: en-core-web-lg NOT needed — presidio uses its own models
